refactor: 🛠  moved files into seperate folders for better structure
