+++ src/components/NoteEditor/utils/pasteHandler.ts
