+++ src/components/NotesFolder/Note/ConfirmNoteDeletion.tsx
