+++ src/components/NotesFolder/utils/filterNotes.ts
